# Make it possible to turn off subcommands that include dependencies