--- a/Cargo.toml
+++ b/Cargo.toml
@@ -59,7 +59,7 @@ serde_json = { version = "1.0.30", features = ["raw_value"] }
 shell-escape = "0.1.4"
 strip-ansi-escapes = "0.1.0"
 tar = { version = "0.4.26", default-features = false }
-tempfile = "3.0"
+tempfile = { version = "3.0", optional = true }
 termcolor = "1.0"
 toml = "0.5.3"
 unicode-xid = "0.2.0"
@@ -113,6 +113,8 @@ test = false
 doc = false
 
 [features]
+default = ["op-install"]
 deny-warnings = []
 vendored-openssl = ["openssl/vendored"]
 pretty-env-logger = ["pretty_env_logger"]
+op-install = ["tempfile"]