--- a/Cargo.toml
+++ b/Cargo.toml
@@ -49,7 +49,7 @@ memchr = "2.1.3"
 num_cpus = "1.0"
 opener = { version = "0.4", optional = true }
 percent-encoding = "2.0"
-rustfix = "0.5.0"
+rustfix = { version = "0.5.0", optional = true }
 same-file = "1"
 semver = { version = "0.10", features = ["serde"] }
 serde = { version = "1.0.82", features = ["derive"] }
@@ -112,9 +112,10 @@ test = false
 doc = false
 
 [features]
-default = ["op-install", "op-doc-open"]
+default = ["op-install", "op-doc-open", "op-fix"]
 deny-warnings = []
 vendored-openssl = ["openssl/vendored"]
 pretty-env-logger = ["pretty_env_logger"]
 op-install = ["tempfile"]
 op-doc-open = ["opener"]
+op-fix = ["rustfix"]