--- a/Cargo.toml
+++ b/Cargo.toml
@@ -47,7 +47,7 @@ log = "0.4.6"
 libgit2-sys = "0.12.7"
 memchr = "2.1.3"
 num_cpus = "1.0"
-opener = "0.4"
+opener = { version = "0.4", optional = true }
 percent-encoding = "2.0"
 remove_dir_all = "0.5.2"
 rustfix = "0.5.0"
@@ -113,8 +113,9 @@ test = false
 doc = false
 
 [features]
-default = ["op-install"]
+default = ["op-install", "op-doc-open"]
 deny-warnings = []
 vendored-openssl = ["openssl/vendored"]
 pretty-env-logger = ["pretty_env_logger"]
 op-install = ["tempfile"]
+op-doc-open = ["opener"]