--- a/src/cargo/ops/mod.rs
+++ b/src/cargo/ops/mod.rs
@@ -12,6 +12,7 @@ pub use self::cargo_generate_lockfile::UpdateOptions;
 pub use self::cargo_install::{install, install_list};
 pub use self::cargo_new::{init, new, NewOptions, VersionControl};
 pub use self::cargo_output_metadata::{output_metadata, ExportInfo, OutputMetadataOptions};
+#[cfg(feature = "op-package-publish")]
 pub use self::cargo_package::{package, PackageOpts};
 pub use self::cargo_pkgid::pkgid;
 pub use self::cargo_read_manifest::{read_package, read_packages};
@@ -21,11 +22,13 @@ pub use self::cargo_uninstall::uninstall;
 #[cfg(feature = "op-fix")]
 pub use self::fix::{fix, fix_maybe_exec_rustc, FixOptions};
 pub use self::lockfile::{load_pkg_lockfile, resolve_to_string, write_pkg_lockfile};
+#[cfg(feature = "op-package-publish")]
+pub use self::registry::publish;
 pub use self::registry::HttpTimeout;
 pub use self::registry::{configure_http_handle, http_handle_and_timeout};
 pub use self::registry::{http_handle, needs_custom_http_transport, registry_login, search};
 pub use self::registry::{modify_owners, yank, OwnersOptions, PublishOpts};
-pub use self::registry::{publish, registry_configuration, RegistryConfig};
+pub use self::registry::{registry_configuration, RegistryConfig};
 pub use self::resolve::{
     add_overrides, get_resolved_packages, resolve_with_previous, resolve_ws, resolve_ws_with_opts,
 };
@@ -40,6 +43,7 @@ mod cargo_generate_lockfile;
 mod cargo_install;
 mod cargo_new;
 mod cargo_output_metadata;
+#[cfg(feature = "op-package-publish")]
 mod cargo_package;
 mod cargo_pkgid;
 mod cargo_read_manifest;