There are some dependencies of the cargo crate that are only used by one or two cargo subcommands. Various tools that use cargo from crates.io don't need most of cargo's built in subcommands, but only a subset. For example, cargo-outdated only does resolving, and cargo-udeps only does `cargo check` with some extras. 

This PR adds default-on cargo features for certain subcommands that pull in dependencies. This allows custom cargo subcommands to have a smaller footprint of transitive dependencies and faster build time.

I ran some tests in this repo with `cargo +nightly build` first with the state prior to this PR and then with `--default-features` enabled. It showed 5 less crates to compile, as well as 2 seconds in saved build time (57.3s to 55.7s), which mostly stemmed from reduced build time for the `cargo` crate (28.1s to 26.5s). Apparently the 5 crates avoided weren't part of the critical path of the compilation DAG, but this might be different in scenarios with more or less parallelism than I have. Check how the green line goes down much quicker prior to the 16 second mark:

Graph prior to this PR:
![grafik](https://user-images.githubusercontent.com/8872119/85173513-b112bd80-b273-11ea-9ece-aeaf1d8d2303.png)
Graph after this PR with `--no-default-features`:
![grafik](https://user-images.githubusercontent.com/8872119/85173703-1ebee980-b274-11ea-9388-051e1e72bb92.png)