@alexcrichton 2020-06-23
Although checked on CI the downside to this is that this is adding `#[cfg]` to maintain and keep working over time. The wins here don't really seem all that big (those graphs look almost identical) and it's a lot of annotations throughout we'd have to grapple with. I think the idea here is fine, but it appears that at least this iteration doesn't seem to really pull the weight of the implementation.