@est31 2020-06-23
I've done a bunch of measurements now with varying scenarios. Indeed I have apparently been testing the scenario with the smallest impact. The impact is larger for less powerful configurations. Tables for the whole crate graph time:

| description | time without PR | time with PR + no-default-features | reduction seconds | reduction percent |
| - | - | - | - | - |
| debug 8 jobs | 53.8s | 52.7s | 1.1 | 2.0 |
| release 8 jobs | 121.8s | 116.3s | 5.5 |  4.5 |
| release 4 jobs | 149.8s | 142.6s | 7.2 | 4.8 |
| debug 4 jobs | 62.1s | 60.1s | 2.0 | 3.2 |

Only cargo crate build time:

| description | time without PR | time with PR + no-default-features | reduction seconds | reduction percent |
| - | - | - | - | - |
| debug 8 jobs | 27.18s | 25.36s | 1.82 | 6.6 | 
| release 8 jobs | 58.39s | 53.58s | 4.81 | 8.2 |
| release 4 jobs | 64.76s | 59.27s | 5.49 | 8.5 |
| debug 4 jobs | 27.3s | 25.71s |  1.59 | 5.8 |

Up to 7 saved seconds, how does that sound?