@est31 2020-06-23
Regarding the concern about the added `#[cfg(feature = "...")]` statements, I think most of them are quite straightforward, next to modules that concern the subcommands. The only place where they are a bit annoying is in the `common_for_install_and_uninstall.rs` file, which mainly contains stuff needed by `cargo install`. What do you think about moving stuff that only cargo install needs into the install-specific file? Would that resolve your concerns?