@alexcrichton 2020-06-24
TBH the whole crate builds look like the differences are just noise in build times, and the cargo-crate-itself doesn't seem too relevant since it seems like the build time for Cargo only matters when you're building it as a dependency, where it's either never changing or you're building the whole thing from scratch.

The goal here seems to be to reuse Cargo's functionality in external crates. I think that goal is probably best served by splitting Cargo up rather than trying to remove pieces of Cargo that aren't needed. That's both more maintainable for the authors of Cargo and clearer to see what depends on what. It's a much more significant undertaking, though.