@est31 2020-06-24
> differences are just noise

I ran the comparison multiple times using multiple scenarios and each time there is a noticeable single digit percent improvement. If it were just noise, there would have been cases where it would be slower or faster on a random basis. I didn't observe that. I invite you to do `cargo +nightly build --release -j 4 --no-default-features -Ztimings` yourself on the commit this PR bases on vs the last commit in this PR and check whether it's really 7 seconds shorter. I ran `cargo update` since I posted the comment with the table above, and now the difference in that specific scenario is even larger, 9 seconds.

For rustc, performance improvement PRs are merged all the time that add single digit percent improvments in compile time, and it's great that this happens. It's the cumulation of multiple such PRs that creates a noticeable "wow" improvement in compile time. See below, I'm already testing how splitting up cargo speeds up the build.

> the cargo-crate-itself doesn't seem too relevant

I included it because of the possible impact on the development workflow.

> best served by splitting Cargo up rather than trying to remove pieces of Cargo that aren't needed.

IMO ideally it shouldn't be "either or" but "both" :). How would you put subcommands into separate crates so that users of cargo can disable them one by one? There is a lot of subcommands, you'd create lots of very tiny crates. I think for them the optimal solution is to have one crate that contains code for all subcommands and then you have cargo features to turn off demanding subcommands (basically what this PR is doing).

Actually I have done some experiments already, building upon this PR's last commit to do right that, and split up cargo even further ([reddit comment](https://www.reddit.com/r/rust/comments/he5ran/rusts_huge_compilation_units/fvq2z6u/)). But the wall clock improvements are cumulative, but in a similar ballpark to the improvements of this PR, which you said was noise (hopefully i was able to convince you it isn't!). And it increases the total CPU time by far more than the wall clock time improvements. See these screenshots (8 jobs, `--no-default-features` enabled in all of them):

baseline as of this PR (52.7s wall clock time):

![Screenshot_20200624_174536](https://user-images.githubusercontent.com/8872119/85587991-901ae580-b642-11ea-9f39-ddaa27d3bc34.png)

cargo-core/cargo split (50.3s cargo-core finishes a bit faster than the old cargo crate, but the difference is smaller than the added time of the cargo crate that while not affecting wall clock time, does affect total CPU time):

![Screenshot_20200624_174347](https://user-images.githubusercontent.com/8872119/85587675-47fbc300-b642-11ea-83c0-b8d654a0c846.png)

splitting off cargo-util off cargo-core (49.2s ... I agree that this build time improvement is inside the noise interval, improvement might be real but very small):

![Screenshot_20200624_174234](https://user-images.githubusercontent.com/8872119/85587520-239fe680-b642-11ea-9833-9d1bc91880c4.png)