@alexcrichton 2020-06-25
Sorry I don't mean to say the results are reproducible, I mean moreso to say that they're so small they don't feel to me like they'll carry the weight that comes with the changes. I'm well aware of how the compiler is getting faster with tiny changes, but most of those changes are not causing any real change in maintenance burden. There's a lot of `#[cfg]` simply sprinkled around as part of this PR, and that's extremely difficult to maintain over time.

I think your use case is using cargo-as-a-library, right? If so I think that the best ways to solve this right now are:

* Use `cargo metadata` if you can -- or if you can't try adding metadata printed by `cargo metadata` to parse
* Split Cargo into separate crates so you don't have to pull in everything-at-once.

The latter strategy is unfortunately also somewhat tricky because to make Cargo manageable we'll want to make Cargo have its own workspace, but this runs afoul of including Cargo as a submodule into rust-lang/rust where we can't have nested workspaces.

Given that I feel like the best way forward would be to look into seeing if `cargo metadata` can be extended. That way the build time of `cargo` itself doesn't matter too too much.