@est31 2020-06-26
I wonder if there are cfg statements that are more straightforward than the ones added by this PR like

```Rust
#[cfg(feature = "op-package-publish")]
pub fn publish(ws: &Workspace<'_>, opts: &PublishOpts<'_>) -> CargoResult<()> {
```

and

```Rust
#[cfg(feature = "op-fix")]
mod fix;
```

but whatever, I'm giving up trying to convince you that this PR is useful.

Given your reluctance to merge this PR, I doubt I'll invest more time into further improvements, higher hanging fruit etc. I need to be able to trust that my contribution will be valued, that compile time is a priority. This doesn't really induce that trust.

Have a great day.